requirements.txt
Flask==3.0.0
web3==6.11.1
py-solc-x==2.0.2
